Core - More efficient edge attribute access
Make access to a subset of edge attributes faster by using the underlying library's methods rather than generating the array for all edges every time.
This also reduces the memory footprint.

+++ nngt/lib/graph_helpers.py
@@ -34,7 +34,7 @@ from .test_functions import nonstring_container, is_integer
 
 def _edge_prop(prop):
     ''' Return edge property `name` as a distribution dict '''
-    if is_integer(prop) or isinstance(prop, np.float):
+    if is_integer(prop) or isinstance(prop, float):
         return {"distribution": "constant", "value": prop}
     elif isinstance(prop, dict):
         return prop.copy()
